add fusionstor_plugin.py for Migrate
