+++ kvmagent/kvmagent/plugins/vm_plugin.py
@@ -2062,7 +2062,6 @@ class VmPlugin(kvmagent.KvmAgent):
     KVM_VM_CHECK_STATE = "/vm/checkstate"
     KVM_HARDEN_CONSOLE_PATH = "/vm/console/harden"
     KVM_DELETE_CONSOLE_FIREWALL_PATH = "/vm/console/deletefirewall"
-    KVM_FUSIONSTOR_QUERY_PATH = "/fusionstor/query"
 
     VM_OP_START = "start"
     VM_OP_STOP = "stop"
@@ -2438,11 +2437,6 @@ class VmPlugin(kvmagent.KvmAgent):
         return jsonobject.dumps(kvmagent.AgentResponse())
 
     @kvmagent.replyerror
-    def fusionstor_query(self, req):
-        makesure_qemu_with_lichbd()
-        return jsonobject.dumps(kvmagent.AgentResponse())
-
-    @kvmagent.replyerror
     def create_ceph_secret_key(self, req):
         cmd = jsonobject.loads(req[http.REQUEST_BODY])
         sh_cmd = shell.ShellCmd('virsh secret-list | grep %s > /dev/null' % cmd.uuid)
@@ -2496,7 +2490,6 @@ class VmPlugin(kvmagent.KvmAgent):
         http_server.register_async_uri(self.KVM_VM_CHECK_STATE, self.check_vm_state)
         http_server.register_async_uri(self.KVM_HARDEN_CONSOLE_PATH, self.harden_console)
         http_server.register_async_uri(self.KVM_DELETE_CONSOLE_FIREWALL_PATH, self.delete_console_firewall_rule)
-        http_server.register_async_uri(self.KVM_FUSIONSTOR_QUERY_PATH, self.fusionstor_query)
 
         self.register_libvirt_event()
